Rename form class to OrbitalResonancesOfOneMinorPlanet
